+++ adapters/config-format.js
@@ -1,18 +1,17 @@
 const vscode = require('vscode');
 
 var format = function (element) {
-  var show_hosts = vscode.workspace.getConfiguration('smartssh').showHostsInPickLists;
+  var show_hosts = vscode.workspace.getConfiguration('smartssh-smba').showHostsInPickLists;
   var config = {
-    name: (show_hosts) ? element.username + '@' + element.host : element.name, // Used for serverlist
-    username: element.username, // Used for authorization
-    password: element.password, // Used for authorization (can be undefined)
-    host: element.host, // Used for authorization
-    port: element.port, // Used for authorization (can be undefined)
-    privateKey: element.privateKey, // Used for authorization (can be undefined)
-    agent: element.agent, // Used for authorization (can be undefined)
-    project: element.project, // Used for fast button (can be undefined)
-    path: element.path, // Used for `cd` after start session (can be undefined)
-    customCommands: element.customCommands, // Used for specify commands which will execute on session start
+    name: (show_hosts) ? element.username + '@' + element.host : element.name, // 用于服务器列表
+    username: element.username, // 用于授权
+    password: element.password, // 用于授权（可以为undefined）
+    host: element.host, // 用于授权
+    port: element.port, // 用于授权（可以为undefined）
+    privateKey: element.privateKey, // 用于授权（可以为undefined）
+    agent: element.agent, // 用于授权（可以为undefined）
+    customCommands: element.customCommands, // 用于指定会话开始时执行的命令
+    smbMapping: element.smbMapping, // SMB映射配置（本地和远程路径）
   };
   return config;
 };